+++ lib/dnd.dart
@@ -8,6 +8,7 @@ import 'dart:js';
 part 'src/draggable.dart';
 part 'src/draggable_avatar.dart';
 part 'src/draggable_dispatch.dart';
+part 'src/draggable_manager.dart';
 part 'src/dropzone.dart';
 part 'src/dropzone_acceptor.dart';