+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## Version 0.1.1 (?)
 
+* Support for IE10 and IE11 touch screens through pointer events.
 * Removed `disableTouch` and `disableMouse` options. The goal was to unify
   touch and mouse dragging, so it should not be necessary to disable 
   one or the other.